Fix build failure on OSX
Add some conditionals for darwin
