+++ execution/executors/oci/console.go
@@ -1,10 +1,8 @@
 package oci
 
 import (
-	"fmt"
 	"os"
 	"syscall"
-	"unsafe"
 )
 
 // newConsole returns an initialized console that can be used within a container by copying bytes
@@ -28,27 +26,4 @@ func newConsole(uid, gid int) (*os.File, string, error) {
 		return nil, "", err
 	}
 	return master, console, nil
-}
-
-func ioctl(fd uintptr, flag, data uintptr) error {
-	if _, _, err := syscall.Syscall(syscall.SYS_IOCTL, fd, flag, data); err != 0 {
-		return err
-	}
-	return nil
-}
-
-// unlockpt unlocks the slave pseudoterminal device corresponding to the master pseudoterminal referred to by f.
-// unlockpt should be called before opening the slave side of a pty.
-func unlockpt(f *os.File) error {
-	var u int32
-	return ioctl(f.Fd(), syscall.TIOCSPTLCK, uintptr(unsafe.Pointer(&u)))
-}
-
-// ptsname retrieves the name of the first available pts for the given master.
-func ptsname(f *os.File) (string, error) {
-	var n int32
-	if err := ioctl(f.Fd(), syscall.TIOCGPTN, uintptr(unsafe.Pointer(&n))); err != nil {
-		return "", err
-	}
-	return fmt.Sprintf("/dev/pts/%d", n), nil
 }
