+++ execution/executors/oci/reaper.go
@@ -1,3 +1,5 @@
+// +build !darwin
+
 package oci
 
 import (
